+++ src/map_io/map_players_view_packet.cc
@@ -351,7 +351,7 @@ void MapPlayersViewPacket::read(FileSystem& fs,
 									map_object_descr = nullptr;
 						} else
 							map_object_descr = nullptr;
-						f_player_field.map_object_descr[TriangleIndex::None] = map_object_descr;
+						f_player_field.map_object_descr[static_cast<int>(TriangleIndex::None)] = map_object_descr;
 					}
 
 					{  //  triangles
@@ -554,7 +554,7 @@ void MapPlayersViewPacket::read(FileSystem& fs,
 					}
 					MapObjectData mod = read_unseen_immovable(
 					   egbase, imm_kind, node_immovables_file, node_immovables_file_version);
-					f_player_field.map_object_descr[TriangleIndex::None] = mod.map_object_descr;
+					f_player_field.map_object_descr[static_cast<int>(TriangleIndex::None)] = mod.map_object_descr;
 					f_player_field.constructionsite = mod.csi;
 
 					// Read in whether this field had a border the last time it was seen
@@ -593,7 +593,7 @@ void MapPlayersViewPacket::read(FileSystem& fs,
 								map_object_descr = nullptr;
 					} else
 						map_object_descr = nullptr;
-					f_player_field.map_object_descr[TriangleIndex::None] = map_object_descr;
+					f_player_field.map_object_descr[static_cast<int>(TriangleIndex::None)] = map_object_descr;
 					break;
 				}
 
@@ -603,7 +603,7 @@ void MapPlayersViewPacket::read(FileSystem& fs,
 					//  information about the triangle has not been saved. Fill in
 					//  the information from the game state.
 					f_player_field.terrains.d = f.field->terrain_d();
-					f_player_field.map_object_descr[TriangleIndex::D] = nullptr;
+					f_player_field.map_object_descr[static_cast<int>(TriangleIndex::D)] = nullptr;
 				} else if (f_everseen | bl_everseen | br_everseen) {
 					//  The player has seen the D triangle but does not see it now.
 					//  Load his information about the triangle from file.
@@ -623,14 +623,15 @@ void MapPlayersViewPacket::read(FileSystem& fs,
 					}
 					MapObjectData mod = read_unseen_immovable(
 					   egbase, im_kind, triangle_immovables_file, triangle_immovables_file_version);
-					f_player_field.map_object_descr[TriangleIndex::D] = mod.map_object_descr;
+					f_player_field.map_object_descr[static_cast<int>(TriangleIndex::D)] =
+					   mod.map_object_descr;
 				}
 				if (f_seen | br_seen | r_seen) {
 					//  The player currently sees the R triangle. Therefore his
 					//  information about the triangle has not been saved. Fill in
 					//  the information from the game state.
 					f_player_field.terrains.r = f.field->terrain_r();
-					f_player_field.map_object_descr[TriangleIndex::R] = nullptr;
+					f_player_field.map_object_descr[static_cast<int>(TriangleIndex::R)] = nullptr;
 				} else if (f_everseen | br_everseen | r_everseen) {
 					//  The player has seen the R triangle but does not see it now.
 					//  Load his information about the triangle from file.
@@ -650,7 +651,8 @@ void MapPlayersViewPacket::read(FileSystem& fs,
 					}
 					MapObjectData mod = read_unseen_immovable(
 					   egbase, im_kind, triangle_immovables_file, triangle_immovables_file_version);
-					f_player_field.map_object_descr[TriangleIndex::R] = mod.map_object_descr;
+					f_player_field.map_object_descr[static_cast<int>(TriangleIndex::R)] =
+					   mod.map_object_descr;
 				}
 
 				{  //  edges
@@ -716,7 +718,8 @@ void MapPlayersViewPacket::read(FileSystem& fs,
 							                            kCurrentPacketVersionSurveyAmounts);
 						}
 						try {
-							f_player_field.time_triangle_last_surveyed[TriangleIndex::D] =
+							f_player_field
+							   .time_triangle_last_surveyed[static_cast<int>(TriangleIndex::D)] =
 							   survey_times_file.unsigned_32();
 						} catch (const FileRead::FileBoundaryExceeded&) {
 							throw GameDataError(
@@ -750,7 +753,8 @@ void MapPlayersViewPacket::read(FileSystem& fs,
 							                            kCurrentPacketVersionSurveyAmounts);
 						}
 						try {
-							f_player_field.time_triangle_last_surveyed[TriangleIndex::R] =
+							f_player_field
+							   .time_triangle_last_surveyed[static_cast<int>(TriangleIndex::R)] =
 							   survey_times_file.unsigned_32();
 						} catch (const FileRead::FileBoundaryExceeded&) {
 							throw GameDataError(
@@ -895,7 +899,8 @@ void MapPlayersViewPacket::write(FileSystem& fs, EditorGameBase& egbase, MapObje
 						assert(f_player_field.owner < 0x20);
 						owners_file.unsigned_8(f_player_field.owner);
 						MapObjectData mod;
-						mod.map_object_descr = f_player_field.map_object_descr[TriangleIndex::None];
+						mod.map_object_descr =
+						   f_player_field.map_object_descr[static_cast<int>(TriangleIndex::None)];
 						mod.csi = f_player_field.constructionsite;
 						write_unseen_immovable(&mod, node_immovable_kinds_file, node_immovables_file);
 
@@ -915,7 +920,8 @@ void MapPlayersViewPacket::write(FileSystem& fs, EditorGameBase& egbase, MapObje
 					   (!bl_seen & !br_seen & (f_everseen | bl_everseen | br_everseen)) {
 						terrains_file.unsigned_8(f_player_field.terrains.d);
 						MapObjectData mod;
-						mod.map_object_descr = f_player_field.map_object_descr[TriangleIndex::D];
+						mod.map_object_descr =
+						   f_player_field.map_object_descr[static_cast<int>(TriangleIndex::D)];
 						write_unseen_immovable(
 						   &mod, triangle_immovable_kinds_file, triangle_immovables_file);
 					}
@@ -925,7 +931,8 @@ void MapPlayersViewPacket::write(FileSystem& fs, EditorGameBase& egbase, MapObje
 					   (!br_seen & !r_seen & (f_everseen | br_everseen | r_everseen)) {
 						terrains_file.unsigned_8(f_player_field.terrains.r);
 						MapObjectData mod;
-						mod.map_object_descr = f_player_field.map_object_descr[TriangleIndex::R];
+						mod.map_object_descr =
+						   f_player_field.map_object_descr[static_cast<int>(TriangleIndex::R)];
 						write_unseen_immovable(
 						   &mod, triangle_immovable_kinds_file, triangle_immovables_file);
 					}
@@ -942,7 +949,7 @@ void MapPlayersViewPacket::write(FileSystem& fs, EditorGameBase& egbase, MapObje
 				//  geologic survey
 				if (f_everseen & bl_everseen & br_everseen) {
 					const uint32_t time_last_surveyed =
-					   f_player_field.time_triangle_last_surveyed[TriangleIndex::D];
+					   f_player_field.time_triangle_last_surveyed[static_cast<int>(TriangleIndex::D)];
 					const uint8_t has_info = time_last_surveyed != 0xffffffff;
 					surveys_file.unsigned_8(has_info);
 					if (has_info) {
@@ -952,7 +959,7 @@ void MapPlayersViewPacket::write(FileSystem& fs, EditorGameBase& egbase, MapObje
 				}
 				if (f_everseen & br_everseen & r_everseen) {
 					const uint32_t time_last_surveyed =
-					   f_player_field.time_triangle_last_surveyed[TriangleIndex::R];
+					   f_player_field.time_triangle_last_surveyed[static_cast<int>(TriangleIndex::R)];
 					const uint8_t has_info = time_last_surveyed != 0xffffffff;
 					surveys_file.unsigned_8(has_info);
 					if (has_info) {
